Fixed intermitent crashing with Finally class
Finally was storing a reference to a std::function to execute later. Passing a
lamda to a function wrapped in a std::function using a reference is ok only if
the std::function will be invoked immediately because the std::function is
a temporary value. Now Finally stores a copy of the std::function.